ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
ultralytics==8.0.200
numpy==1.26.2
Pillow==10.1.0
scipy==1.11.4
pyaudio==0.2.13
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
